+++ setup.py
@@ -68,8 +68,6 @@ To create the package for pypi.
    Then push the change with a message 'set dev version'
 """
 
-import os
-
 from setuptools import find_packages, setup
 
 
@@ -99,44 +97,41 @@ EXTRAS["dev"] = []
 for reqs in EXTRAS.values():
     EXTRAS["dev"].extend(reqs)
 
-try:
-    file_path = os.path.dirname(os.path.abspath(__file__))
-    os.symlink(os.path.join(file_path, "examples/scripts"), os.path.join(file_path, "trl/commands/scripts"))
 
-    setup(
-        name="trl",
-        license="Apache 2.0",
-        classifiers=[
-            "Development Status :: 2 - Pre-Alpha",
-            "Intended Audience :: Developers",
-            "Intended Audience :: Science/Research",
-            "License :: OSI Approved :: Apache Software License",
-            "Natural Language :: English",
-            "Operating System :: OS Independent",
-            "Programming Language :: Python :: 3",
-            "Programming Language :: Python :: 3.9",
-            "Programming Language :: Python :: 3.10",
-            "Programming Language :: Python :: 3.11",
-            "Programming Language :: Python :: 3.12",
-        ],
-        url="https://github.com/huggingface/trl",
-        entry_points={
-            "console_scripts": ["trl=trl.commands.cli:main"],
-        },
-        include_package_data=True,
-        package_data={"trl": ["commands/scripts/config/*", "commands/scripts/*", "templates/*.md"]},
-        packages=find_packages(exclude={"tests"}),
-        install_requires=REQUIRED_PKGS,
-        extras_require=EXTRAS,
-        python_requires=">=3.9",
-        long_description=open("README.md", encoding="utf-8").read(),
-        long_description_content_type="text/markdown",
-        zip_safe=False,
-        version=__version__,
-        description="Train transformer language models with reinforcement learning.",
-        keywords="ppo, transformers, huggingface, gpt2, language modeling, rlhf",
-        author="Leandro von Werra",
-        author_email="leandro.vonwerra@gmail.com",
-    )
-finally:
-    os.unlink(os.path.join(file_path, "trl/commands/scripts"))
+setup(
+    name="trl",
+    license="Apache 2.0",
+    classifiers=[
+        "Development Status :: 2 - Pre-Alpha",
+        "Intended Audience :: Developers",
+        "Intended Audience :: Science/Research",
+        "License :: OSI Approved :: Apache Software License",
+        "Natural Language :: English",
+        "Operating System :: OS Independent",
+        "Programming Language :: Python :: 3",
+        "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
+        "Programming Language :: Python :: 3.12",
+    ],
+    url="https://github.com/huggingface/trl",
+    entry_points={
+        "console_scripts": ["trl=trl.cli:main"],
+    },
+    include_package_data=True,
+    package_data={
+        "trl": ["templates/*.md"],
+    },
+    packages=find_packages(exclude={"tests", "tests.slow"}),
+    install_requires=REQUIRED_PKGS,
+    extras_require=EXTRAS,
+    python_requires=">=3.9",
+    long_description=open("README.md", encoding="utf-8").read(),
+    long_description_content_type="text/markdown",
+    zip_safe=False,
+    version=__version__,
+    description="Train transformer language models with reinforcement learning.",
+    keywords="ppo, transformers, huggingface, gpt2, language modeling, rlhf",
+    author="Leandro von Werra",
+    author_email="leandro.vonwerra@gmail.com",
+)
