+++ tests/test_cli.py
@@ -12,34 +12,41 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import subprocess
-import sys
+
+import tempfile
 import unittest
+from io import StringIO
+from unittest.mock import patch
+
+from trl.cli import main
 
 
-class CLITester(unittest.TestCase):
-    @unittest.skipIf(sys.platform.startswith("win"), "Skipping on Windows")
-    def test_sft_cli(self):
-        try:
-            subprocess.run(
-                "trl sft --max_steps 1 --output_dir tmp-sft --model_name_or_path trl-internal-testing/tiny-Qwen2ForCausalLM-2.5 --dataset_name stanfordnlp/imdb --learning_rate 1e-4 --lr_scheduler_type cosine",
-                shell=True,
-                check=True,
-            )
-        except BaseException:
-            self.fail("An error occurred while running the CLI, please double check")
+class TestCLI(unittest.TestCase):
+    def test_dpo(self):
+        with tempfile.TemporaryDirectory() as tmp_dir:  # Create a temporary directory
+            command = f"trl dpo --output_dir {tmp_dir} --model_name_or_path trl-internal-testing/tiny-Qwen2ForCausalLM-2.5 --dataset_name trl-internal-testing/zen --dataset_config standard_preference --report_to none"
+            with patch("sys.argv", command.split(" ")):
+                main()
 
-    @unittest.skipIf(sys.platform.startswith("win"), "Skipping on Windows")
-    def test_dpo_cli(self):
-        try:
-            subprocess.run(
-                "trl dpo --max_steps 1 --output_dir tmp-dpo --model_name_or_path trl-internal-testing/tiny-Qwen2ForCausalLM-2.5 --dataset_name trl-internal-testing/tiny-ultrafeedback-binarized --learning_rate 1e-4 --lr_scheduler_type cosine",
-                shell=True,
-                check=True,
-            )
-        except BaseException:
-            self.fail("An error occurred while running the CLI, please double check")
+    @patch("sys.stdout", new_callable=StringIO)
+    def test_env(self, mock_stdout):
+        command = "trl env"
+        with patch("sys.argv", command.split(" ")):
+            main()
+        self.assertIn("TRL version: ", mock_stdout.getvalue().strip())
 
-    def test_env_cli(self):
-        output = subprocess.run("trl env", capture_output=True, text=True, shell=True, check=True)
-        self.assertIn("- Python version: ", output.stdout)
+    def test_kto(self):
+        with tempfile.TemporaryDirectory() as tmp_dir:  # Create a temporary directory
+            command = f"trl kto --output_dir {tmp_dir} --model_name_or_path trl-internal-testing/tiny-Qwen2ForCausalLM-2.5 --dataset_name trl-internal-testing/zen --dataset_config standard_unpaired_preference --report_to none"
+            with patch("sys.argv", command.split(" ")):
+                main()
+
+    def test_sft(self):
+        with tempfile.TemporaryDirectory() as tmp_dir:  # Create a temporary directory
+            command = f"trl sft --output_dir {tmp_dir} --model_name_or_path trl-internal-testing/tiny-Qwen2ForCausalLM-2.5 --dataset_name trl-internal-testing/zen --dataset_config standard_language_modeling --report_to none"
+            with patch("sys.argv", command.split(" ")):
+                main()
+
+
+if __name__ == "__main__":
+    unittest.main()
